refactor: attendance-status 코드 파일 분리
attendance-status 코드 파일 분리
